+++ model_navigator/api/utilities.py
@@ -14,9 +14,22 @@
 # noqa: D104
 """Public utilities for the Model Navigator API."""
 
-from typing import Callable
+import json
+import logging
+import pathlib
+import tempfile
+from typing import Any, Callable, Optional
 
-from model_navigator.api.config import SizedDataLoader
+from model_navigator.api.config import Framework, OptimizationProfile, SizedDataLoader
+from model_navigator.commands.find_max_batch_size.find_max_batch_size import MaxBatchSizeFinder
+from model_navigator.core.tensor import FRAMEWORK_TO_TENSOR_TYPE, PyTreeMetadata, TensorMetadata
+from model_navigator.exceptions import ModelNavigatorProfilingError
+from model_navigator.runners.registry import get_runner
+from model_navigator.utils.dataloader import to_numpy
+
+logger_name = "model_navigator.api.utilities"
+LOGGER = logging.getLogger(logger_name)
+logging.basicConfig(level=logging.INFO)
 
 
 class UnpackedDataloader:
@@ -48,3 +61,78 @@ class UnpackedDataloader:
         """Return an iterator over the samples in the dataloader with the unpack_fn applied."""
         for sample in self._dataloader:
             yield self._unpack_fn(sample)
+
+
+def find_max_batch_size_till_oom(
+    framework: Framework,
+    model: Any,
+    dataloader: SizedDataLoader,
+    batch_dim: int = 0,
+    max_batch_size_search_limit: Optional[int] = None,
+):
+    """Find the maximum batch size for a model.
+
+    Search is performed by running the model on the dataloader until an OOM error is encountered.
+
+    Args:
+        framework: The framework of the model.
+        model: The model.
+        dataloader: A SizedDataLoader.
+        batch_dim: The batch dimension of the model.
+        max_batch_size_search_limit: Limit the search for the maximum batch size to this value.
+    """
+    if framework == Framework.TORCH:
+        runner_name = "TorchCUDA"
+    elif framework == Framework.TENSORFLOW:
+        runner_name = "TensorFlowCUDA"
+    elif framework == Framework.ONNX:
+        runner_name = "OnnxCUDA"
+    elif framework == Framework.JAX:
+        runner_name = "Jax"
+
+    sample = next(iter(dataloader))
+
+    pytree_metadata = PyTreeMetadata.from_sample(
+        sample=sample, tensor_type=FRAMEWORK_TO_TENSOR_TYPE[framework], prefix="input"
+    )
+    input_metadata = TensorMetadata(pytree_metadata=pytree_metadata)
+
+    profiling_sample = input_metadata.flatten_sample(sample=sample)
+    profiling_sample = {name: to_numpy(tensor, from_framework=framework) for name, tensor in profiling_sample.items()}
+
+    for name, tensor in profiling_sample.items():
+        shape = list(tensor.shape)
+        shape[batch_dim] = -1
+        input_metadata.add(name=name, shape=shape, dtype=tensor.dtype)
+
+    with tempfile.NamedTemporaryFile() as temp_file:
+        results_path = pathlib.Path(temp_file.name)
+
+        runner = get_runner(runner_name)(
+            model=model,
+            input_metadata=input_metadata,
+            output_metadata=TensorMetadata(),
+        )  # pytype: disable=not-instantiable
+        try:
+            LOGGER.info("Starting max batch size search.")
+            MaxBatchSizeFinder(
+                profile=OptimizationProfile(max_batch_size=max_batch_size_search_limit),
+                batch_dim=batch_dim,
+                results_path=results_path,
+            ).run(
+                runner=runner,
+                profiling_sample=profiling_sample,
+                sample_id=0,
+            )
+        finally:
+            if results_path.is_file():
+                with open(results_path) as file:
+                    max_bs_line = file.readlines()[-1]
+                    results_dict = json.loads(max_bs_line)
+                    if "batch_size" in results_dict:
+                        LOGGER.info(f"Max batch size: {results_dict['batch_size']}")
+                    else:
+                        raise ModelNavigatorProfilingError("Max batch size not found.")
+            else:
+                raise ModelNavigatorProfilingError("Max batch size not found.")
+        return results_dict["batch_size"]